testdrive: improve documentation and layout
Reorganize the crate's exports to be a bit more sensible, and document
the somewhat unique error handling philosophy.
