+++ src/testdrive/error.rs
@@ -3,12 +3,36 @@
 // This file is part of Materialize. Materialize may not be used or
 // distributed without the express permission of Materialize, Inc.
 
-use atty::Stream;
+//! Error handling.
+//!
+//! Testdrive takes pains to provide better error messages than are typical of a
+//! Rust program. The main error type, [`Error`], is not intentionally not
+//! constructible from from underlying errors without providing additional
+//! context. The [`ResultExt`] trait can be used to ergonomically add this
+//! context to the error within the result. For example, here is an idiomatic
+//! example of handling a filesystem error:
+//!
+//! ```rust
+//! use std::fs::File;
+//! use std::io::Read;
+//! use testdrive::error::{Error, ResultExt};
+//!
+//! fn check_file(path: &str) -> Result<bool, Error> {
+//!     let mut file = File::open(&path).err_ctx(format!("opening {}", path))?;
+//!
+//!     let mut contents = String::new();
+//!     file.read_to_string(&mut contents).err_ctx(format!("reading {}", path))?;
+//!     Ok(contents.contains("AOK"))
+//! }
+//! ```
+
 use std::error::Error as StdError;
 use std::fmt;
 use std::fmt::Write as FmtWrite;
 use std::io;
 use std::io::Write;
+
+use atty::Stream;
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
 #[derive(Debug)]
@@ -19,7 +43,7 @@ pub enum Error {
     },
     General {
         ctx: String,
-        cause: Box<dyn StdError>,
+        cause: Option<Box<dyn StdError>>,
         hints: Vec<String>,
     },
     Usage {
@@ -62,7 +86,10 @@ impl Error {
             Error::General { ctx, cause, hints } => {
                 let color_spec = ColorSpec::new();
                 write_error_heading(&mut stderr, &color_spec)?;
-                writeln!(&mut stderr, "{}: {}", ctx, cause)?;
+                write!(&mut stderr, "{}", ctx)?;
+                if let Some(cause) = cause {
+                    writeln!(&mut stderr, ": {}", cause)?;
+                }
                 for hint in hints {
                     stderr.set_color(&color_spec.clone().set_bold(true))?;
                     write!(&mut stderr, " hint: ")?;
@@ -147,7 +174,13 @@ impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match self {
             Error::Input { err, .. } => write!(f, "{}", err.msg),
-            Error::General { cause, .. } => cause.fmt(f),
+            Error::General { ctx, cause, .. } => {
+                write!(f, "{}", ctx)?;
+                if let Some(cause) = cause {
+                    write!(f, ": {}", cause)?;
+                }
+                Ok(())
+            }
             Error::Usage { details, .. } => write!(f, "usage error: {}", details),
         }
     }
@@ -189,14 +222,8 @@ where
     {
         self.map_err(|err| Error::General {
             ctx,
-            cause: Box::new(err),
+            cause: Some(Box::new(err)),
             hints: Vec::new(),
         })
-    }
-}
-
-impl From<Error> for String {
-    fn from(e: Error) -> String {
-        e.to_string()
     }
 }
